Optimisation and Mobile Detection Improvements
Optimised item menu selection and improved mobile detection.
